aalc: add sensor level 3 and event log
Summary:
- modify fan table
- block auto tune command util bic is ready
- disable auto tune will not set duty to 0
- add more error log event
- add sensor COOLING_CAPACITY
- add sensor BPB_RACK_LEVEL_3
- extend COOLING_CAPACITY modbus command to 2 register
- add modbus shell command for test
- modify the threshold of y_filter to be dynamic
- fix threshold struct member last_status to the correct size

Test Plan:
- Build code: PASS

+++ meta-facebook/aalc-rpu/src/platform/plat_threshold.h
@@ -28,7 +28,7 @@ typedef struct {
 	uint32_t arg0;
 
 	// priv data
-	uint8_t last_status; // record the last status
+	uint32_t last_status; // record the last status
 	uint32_t last_value; // record the last value
 } sensor_threshold;
